requirements_ryan.txt
-e git+ssh://git@github.com/AntoineLuciano/ABC-SBI@e9cfb878d1d47ee92d52e381de04cbdf375a2055#egg=abcnre&subdirectory=abcnre
absl-py==2.1.0
altair==5.5.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arviz==0.21.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
autograd==1.7.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dask==2025.2.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
deneb==1.0.1
diffeqtorch==1.0.0
distributed==2025.2.0
elfi==0.8.7
etils==1.12.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
flatlatex==0.15
flax==0.10.4
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
GPy==1.13.2
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5netcdf==1.5.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipyparallel==9.0.1
ipython==8.33.0
ipywidgets==8.1.5
isoduration==20.11.0
jabbar==0.0.16
jax==0.5.2
jaxlib==0.5.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julia==0.6.2
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.29.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nflows==0.14
notebook==7.3.2
notebook_shim==0.2.4
numdifftools==0.9.41
numpy==1.25.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramz==0.9.6
parso==0.8.4
partd==1.4.2
peppercorn==0.6
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
portpicker==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==6.30.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyabc==0.12.15
pyabcranger==0.0.72
pycparser==2.22
Pygments==2.19.1
pyknos==0.16.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
pyro-api==0.1.2
pyro-ppl==1.9.1
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
-e git+ssh://git@github.com/AntoineLuciano/ABC-SBI@e9cfb878d1d47ee92d52e381de04cbdf375a2055#egg=sampleproject&subdirectory=abcnre
sbi==0.21.0
sbibm==1.1.0
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
sympy==1.13.1
tblib==3.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorstore==0.1.72
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.9
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
vega-datasets==0.9.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wsproto==1.2.0
xarray==2025.1.2
xarray-einstats==0.8.0
zict==3.0.0
zipp==3.21.0
